fix(prompt): improve error message logging
I am changing the way error messages are logged to be more informative.
Previously, the error message was only included in the description field of the
prompt result. Now, I am also including the error message in the messages field
of the prompt result, which will make it easier to debug issues.